Adds recipe_ingredients factory

+++ db/migrate/20190926171334_create_recipe_ingredients.rb
@@ -3,7 +3,7 @@ class CreateRecipeIngredients < ActiveRecord::Migration[5.2]
     create_table :recipe_ingredients do |t|
       t.integer :measurement_type
       t.integer :measurement
-      t.integer :quantity
+      t.decimal :quantity
       t.references :ingredient, foreign_key: true
 
       t.timestamps
